move customizer to service discovery dir add host port customizer add licence
