+++ registry/servicediscovery/customizer/metadata_service_url_params_customizer.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package event
+package customizer
 
 import (
 	"encoding/json"
