+++ registry/servicediscovery/customizer/protocol_ports_metadata_customizer.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package event
+package customizer
 
 import (
 	"encoding/json"
